Assignment_portal/requirements.txt
bcrypt==4.2.0
blinker==1.9.0
click==8.1.7
colorama==0.4.6
dnspython==2.7.0
Flask==3.1.0
Flask-JWT-Extended==4.6.0
MarkupSafe==3.0.2
PyJWT==2.9.0
pymongo==4.10.1
Werkzeug==3.1.3
